xfs_copy: fix copy of hard 4k devices
If we have a pure 4k device with no 512 emulation, xfs_copy
fails straightaway because it tries to do a 512-byte direct
IO read of the superblock.

Do like we do in xfs_db, and read in the max possible sector size,
because we don't yet know what the filesystem's sector size is.

This fixes a failure in xfs/032 on a hard 4k device.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Brian Foster <bfoster@redhat.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Dave Chinner <david@fromorbit.com>