xfsprogs: blkid is now mandatory
Because blkid is here for a long time, I hereby propose a patch for
removing support for NOT having blkid. The current support through
set of #ifdef is prone to errors like
making a patch just in one of the branches, and according to a
recent talk between Christoph and Eric, it is not necessary to keep
it supported.

Remove code for checking ENABLE_BLKID, and the code when
ENABLE_BLKID is not defined.  The only use of libdisk was in the
removed code, so remove libdisk too.  It makes blkid required for
compilation.

[dchinner: also remove include/volume.h and include/dvh.h as
 suggested by Christoph during review. ]

Signed-off-by: Jan Tulak <jtulak@redhat.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Dave Chinner <david@fromorbit.com>